Representar el dinero del dashboard con BigDecimal
Las dos sumas del dashboard devolvian Double y el DTO lo exponia como
Double, mientras que todo el resto del sistema maneja importes con
BigDecimal. Ademas de la inconsistencia, Double pierde centavos en
cuanto los montos crecen. Pasan a BigDecimal en el repositorio, el
servicio y la respuesta.

+++ backend/src/main/java/com/uade/tpejemplo/dto/response/DashboardStatsResponse.java
@@ -3,6 +3,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 @Data
 @NoArgsConstructor
@@ -10,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class DashboardStatsResponse {
     private Long cantidadClientes;
     private Long cantidadCreditos;
-    private Double montoTotalFinanciado;
-    private Double montoTotalCobrado;
+    private BigDecimal montoTotalFinanciado;
+    private BigDecimal montoTotalCobrado;
 }
